Move fingerprintingConvertCommand, update type
